Move begin from MetaPhysicL to MetaPhysicL::detail
In nhlomann json there is this code
```
using std::begin;
begin(libMesh::VectorValue<MetaPhysicL::DualNumber>>)
```
and because of the nested MetaPhysicL type the compiler considers
`MetaPhysicL::begin` a valid overload to consider.
`MetaPhysicL::begin` instantiates `std::begin` for which there
is no such function for libMesh::VectorValue, which leads to
a compilation error.